add typecast to agree with return val

+++ tim.c
@@ -78,32 +78,32 @@ uint16_t get_atim_capval(struct adv_tim* atim_ptr, uint8_t reg)
 	switch(reg)
 	{
 		case ATIM_CC_REG_1:
-			return atim_ptr->CCR1;
+			return (uint16_t)atim_ptr->CCR1;
 			break;
 
 		case ATIM_CC_REG_2:
-			return atim_ptr->CCR2;
+			return (uint16_t)atim_ptr->CCR2;
 			break;
 
 		case ATIM_CC_REG_3:
 
-			return atim_ptr->CCR3;
+			return (uint16_t)atim_ptr->CCR3;
 			break;
 
 		case ATIM_CC_REG_4:
-			return atim_ptr->CCR4;
+			return (uint16_t)atim_ptr->CCR4;
 			break;
 
 		case ATIM_CC_REG_5:
-			return atim_ptr->CCR5;
+			return (uint16_t)atim_ptr->CCR5;
 			break;
 
 		case ATIM_CC_REG_6:
-			return atim_ptr->CCR6;
+			return (uint16_t)atim_ptr->CCR6;
 			break;
 
 		default:
-			return atim_ptr->CCR1;
+			return (uint16_t)atim_ptr->CCR1;
 			break;
 	}
